requirements.txt
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
psycopg2-binary==2.9.11
PySocks==1.7.1
python-dotenv==1.2.1
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.1
